Extract the online game functionality to a custom hook
